+++ cnccode/customClasses/SDManagerDashboard/ServiceRequestSummaryDTO.php
@@ -148,8 +148,10 @@ class ServiceRequestSummaryDTO implements \JsonSerializable
     private $slaFixHoursP2;
     /** @var float */
     private $slaFixHoursP3;
-
-
+    /** @var string */
+    private $contactName;
+        /** @var string */
+    private $emailSubjectSummary;
     /**
      * ServiceRequestSummaryDTO constructor.
      */
@@ -237,8 +239,11 @@ class ServiceRequestSummaryDTO implements \JsonSerializable
         $instance->slaP3PenaltiesAgreed       = $dbeCustomer->getValue(\DBECustomer::slaP3PenaltiesAgreed);
         $instance->slaFixHoursP1              = $dbeCustomer->getValue(\DBECustomer::slaFixHoursP1);
         $instance->slaFixHoursP2              = $dbeCustomer->getValue(\DBECustomer::slaFixHoursP2);
-        $instance->slaFixHoursP3              = $dbeCustomer->getValue(\DBECustomer::slaFixHoursP3);
-        return $instance;
+        $instance->slaFixHoursP3              = $dbeCustomer->getValue(\DBECustomer::slaFixHoursP3);        
+        $instance->contactName                = $problem->getValue(DBEJProblem::contactName);
+        $instance->emailSubjectSummary                = $problem->getValue(DBEJProblem::emailSubjectSummary);
+
+        return $instance;        
     }
 
     private static function getProblemHistoryURL($problemId)
@@ -293,7 +298,9 @@ class ServiceRequestSummaryDTO implements \JsonSerializable
             "slaP3PenaltiesAgreed"       => $this->slaP3PenaltiesAgreed,
             "slaFixHoursP1"              => $this->slaFixHoursP1,
             "slaFixHoursP2"              => $this->slaFixHoursP2,
-            "slaFixHoursP3"              => $this->slaFixHoursP3,
+            "slaFixHoursP3"              => $this->slaFixHoursP3,   
+            "contactName"                =>$this->contactName,
+            "emailSubjectSummary"        => $this->emailSubjectSummary,
         ];
     }