+++ htdocs/components/CurrentActivityReportComponent/subComponents/InboxEscalationsComponent.js
@@ -63,6 +63,17 @@ class InboxEscalationsComponent extends React.Component {
             },
             {
                 hide: false,
+                order: 11.5,
+                path: "contactName",
+                key: 'contactName',
+                label: "",
+                hdToolTip: "Contact",
+                icon: "fal fa-2x fa-id-card-alt color-gray2 ",
+                sortable: false,
+                hdClassName: "text-center",
+            },
+            {
+                hide: false,
                 order: 5,
                 path: null,
                 label: "",
@@ -120,7 +131,7 @@ class InboxEscalationsComponent extends React.Component {
             {
                 hide: false,
                 order: 12,
-                path: "reason",
+                path: "emailSubjectSummary",
                 label: "",
                 hdToolTip: "Description of the Service Request",
                 icon: "fal fa-2x fa-file-alt color-gray2 ",
@@ -132,8 +143,7 @@ class InboxEscalationsComponent extends React.Component {
                         {
                             className: "pointer",
                             onClick: () => srDescription(problem),
-                            dangerouslySetInnerHTML: {__html: problem.reason}
-
+                            dangerouslySetInnerHTML: {__html: problem.emailSubjectSummary}
                         },
                     ),
             },
@@ -193,7 +203,7 @@ class InboxEscalationsComponent extends React.Component {
         const {data} = this.props;
         return (
             <Fragment>
-                <ServiceRequestSummary data={data}/>
+                <ServiceRequestSummary key={data.serviceRequestId} data={data}/>
                 {getTableElement()}
             </Fragment>
         )