+++ htdocs/components/CurrentActivityReportComponent/subComponents/InboxOpenSRComponent.js
@@ -93,6 +93,17 @@ class InboxOpenSRComponent extends React.Component {
             },
             {
                 hide: false,
+                order: 11.5,
+                path: "contactName",
+                key: 'contactName',
+                label: "",
+                hdToolTip: "Contact",
+                icon: "fal fa-2x fa-id-card-alt color-gray2 ",
+                sortable: false,
+                hdClassName: "text-center",
+            },
+            {
+                hide: false,
                 order: 6,
                 path: "problemID",
                 label: "",
@@ -137,7 +148,7 @@ class InboxOpenSRComponent extends React.Component {
             {
                 hide: false,
                 order: 12,
-                path: "reason",
+                path: "emailSubjectSummary",
                 label: "",
                 hdToolTip: "Description of the Service Request",
                 icon: "fal fa-2x fa-file-alt color-gray2 ",
@@ -149,7 +160,7 @@ class InboxOpenSRComponent extends React.Component {
                         {
                             className: "pointer",
                             onClick: () => srDescription(problem),
-                            dangerouslySetInnerHTML: {__html: problem.reason}
+                            dangerouslySetInnerHTML: {__html: problem.emailSubjectSummary}
                         },
                     ),
             },
@@ -247,9 +258,8 @@ class InboxOpenSRComponent extends React.Component {
 
     render() {
         const {getTableElement,} = this;
-        return [
-            getTableElement(),
-        ];
+        return getTableElement();
+
     }
 }