+++ htdocs/components/CurrentActivityReportComponent/subComponents/InboxSalesComponent.js
@@ -80,7 +80,17 @@ class InboxSalesComponent extends React.Component {
                 ],
             },
             ColumnRenderer.getOnHoldColumn(),
-
+            {
+                hide: false,
+                order: 11.5,
+                path: "contactName",
+                key: 'contactName',
+                label: "",
+                hdToolTip: "Contact",
+                icon: "fal fa-2x fa-id-card-alt color-gray2 ",
+                sortable: false,
+                hdClassName: "text-center",
+            },
             ColumnRenderer.getSLABreachedColumn(),
             {
                 hide: false,
@@ -161,7 +171,7 @@ class InboxSalesComponent extends React.Component {
             {
                 hide: false,
                 order: 12,
-                path: "reason",
+                path: "emailSubjectSummary",
                 label: "",
                 hdToolTip: "Description of the Service Request",
                 icon: "fal fa-2x fa-file-alt color-gray2 ",
@@ -173,7 +183,7 @@ class InboxSalesComponent extends React.Component {
                         {
                             className: "pointer",
                             onClick: () => srDescription(problem),
-                            dangerouslySetInnerHTML: {__html: problem.reason}
+                            dangerouslySetInnerHTML: {__html: problem.emailSubjectSummary}
                         },
                     ),
             },